# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VenueVox API Specification",
    "description": "API contract for VenueVox Entertainment Marketplace. Provides endpoints for auth, venues, slots, bookings, chat, and reviews.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Mock Local Development Server"
    }
  ],
  "paths": {
...
    "/slots/discover": {
      "get": {
        "summary": "Search Bookable Slots",
        "parameters": [
          { "name": "location", "in": "query", "schema": { "type": "string" } },
          { "name": "date", "in": "query", "schema": { "type": "string" } },
          { "name": "budget", "in": "query", "schema": { "type": "integer" } },
          { "name": "venueType", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Matching Available calendar slots"
          }
        }
      }
    }
  }
}